requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
clang==16.0.1.1
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==0.7
crypto==1.4.1
cryptography==42.0.5
cssselect2==0.7.0
Django==5.0.3
django-autoslug==1.9.9
django-ckeditor==6.0.0
django-crispy-forms==2.0
django-fontawesome==1.0
django-js-asset==2.2.0
django-paypal==2.0
djangorestframework==3.15.0
git-filter-repo==2.38.0
html5lib==1.1
idna==3.4
lxml==5.2.1
Naked==0.1.32
oscrypto==1.3.0
Pillow==10.0.1
psycopg2==2.9.9
pycparser==2.22
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.4
sslcommerz-lib==1.0
style==1.1.0
svglib==1.5.1
tinycss2==1.2.1
typed-ast==1.5.5
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.2
update==0.0.1
uritools==4.0.2
urllib3==2.0.6
webencodings==0.5.1
xhtml2pdf==0.2.15
